pyproject.toml
[build-system]
requires = ["hatchling>=1.12.0"]
build-backend = "hatchling.build"

[project]
name = "netinfer"
version = "0.3.0"
description = "A Snakemake pipeline for microbiome network inference"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
authors = [
    {name = "Heyu Lin", email = "heyu.lin@qut.edu.au"},
]

dependencies = [
    "snakemake>=9.0.0",
    "pyyaml",
    "pandas",
    "numpy",
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

[project.urls]
Homepage = "https://github.com/SilentGene/netinfer"
Repository = "https://github.com/SilentGene/netinfer"

[project.scripts]
netinfer = "netinfer.cli:main"

[tool.hatch.build]
packages = ["netinfer"]

[tool.hatch.build.targets.wheel]
packages = ["netinfer"]

[tool.hatch.build.targets.sdist]
include = [
    "/netinfer",
    "/test",
    "/*.md",
    "/LICENSE"
]